♻️(malware_detection) add helper to encrypt sensitive data
The usage of celery can leak sensitive data. In the JCOP bcakend we have
to pass to every tasks the api key, this api key will be present in
redis and also logged. We don't want to leak it. So we use the Fernet
algorithm from the cryptography library to encrypt it and then decrypt
it when we need to use it.