♻️(resource-server) make token introspection earlier
The base implementation from Mozilla let the backend in charge
of making the introspection (probably to stay generic), but
in LaSuite project we will always perform introspection when
acting as resource server.

To improve domain consistency and "payload" access to other part
of the code, we chose to make the introspection earlier in the
authentication process.

Note the introspection payload, does not contain much information,
in our current implementation we need it to contain at least the
user's `sub``.